@@ -4,15 +4,15 @@
   "private": true,
   "homepage": "https://zero-to-mastery.github.io/ZtM-Job-Board",
   "dependencies": {
-    "google-map-react": "^1.1.2",
-    "node-sass-chokidar": "^1.3.0",
-    "npm-run-all": "^4.1.3",
-    "react": "^16.3.1",
-    "react-dom": "^16.3.1",
+    "google-map-react": "^1.1.4",
+    "node-sass-chokidar": "^1.3.4",
+    "npm-run-all": "^4.1.5",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "react-favicon": "0.0.14",
-    "react-scripts": "^2.1.3",
+    "react-scripts": "^3.0.0",
     "react-search-input": "^0.11.3",
-    "tachyons": "^4.9.1"
+    "tachyons": "^4.11.1"
   },
   "scripts": {
     "start": "npm-run-all -p watch-css start-js",
@@ -26,7 +26,7 @@
     "deploy": "gh-pages -d build"
   },
   "devDependencies": {
-    "gh-pages": "^1.2.0"
+    "gh-pages": "^2.0.1"
   },
   "browserslist": [
     ">0.2%",
