@@ -1,12 +1,15 @@
 import React from 'react';
 import SearchBox from './SearchBox';
 
-const Search = ({toggleMenu, onSearchChange,  keyPress, responsiveSearch}) => {
-      return(
-            <div className="search-area flex flex-wrap items-center ml4 mt1 z-1">
-              <SearchBox responsiveSearch={responsiveSearch} searchChange={onSearchChange} keyPress={keyPress}/>
-            </div>
-      );
-}
+const Search = ({ onSearchChange, responsiveSearch }) => {
+  return (
+    <div className="search-area flex flex-wrap items-center ml4 mt1 z-1">
+      <SearchBox
+        responsiveSearch={responsiveSearch}
+        searchChange={onSearchChange}
+      />
+    </div>
+  );
+};
 
 export default Search;
