@@ -5,5 +5,5 @@ import App from './App';
 // import registerServiceWorker from './registerServiceWorker';
 import 'tachyons';
 
-ReactDOM.render (<App />,document.getElementById('root'));
+ReactDOM.render(<App />, document.getElementById('root'));
 // registerServiceWorker();
